Reverse dependency between lyra-operator and lyra
This commit adds a lyra build that includes the controller command to
this project so that lyra is used as a library.

This means that there will be two lyra binaries. One that has the
controller command built in, and one that doesn't. We might want to
consider other options.